% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{dd_target}
\alias{dd_target}
\title{The release target triple for the current platform}
\usage{
dd_target()
}
\value{
One of the target triples \code{data-dict} publishes binaries for.
}
\description{
The release target triple for the current platform
}
\examples{
dd_target()
}
